requirements.txt

Pillow==10.1.0
PyAutoGUI==0.9.54
pynput==1.7.6
python-dotenv==1.0.0
python-xlib==0.33
python3-xlib==0.15